Extract retaining context check out
