+++ circuit/src/main/kotlin/com/slack/circuit/retained/Continuity.kt
@@ -15,19 +15,7 @@
  */
 package com.slack.circuit.retained
 
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
 import androidx.lifecycle.ViewModel
-
-internal fun Context.findActivity(): Activity? {
-  var context = this
-  while (context is ContextWrapper) {
-    if (context is Activity) return context
-    context = context.baseContext
-  }
-  return null
-}
 
 class Continuity : ViewModel(), RetainedStateRegistry {
   private val delegate = RetainedStateRegistryImpl()
